feat: add --no-install to make
Add the --no-install option to the make command to build the rock w/o installing it.

+++ src/luarocks/build.lua
@@ -23,6 +23,7 @@ build.opts = util.opts_table("build.opts", {
    verify = "boolean",
    check_lua_versions = "boolean",
    pin = "boolean",
+   no_install = "boolean"
 })
 
 do
@@ -385,16 +386,18 @@ function build.build_rockspec(rockspec, opts)
       return name, version
    end   
 
-   if repos.is_installed(name, version) then
-      repos.delete_version(name, version, opts.deps_mode)
+   if not opts.no_install then
+      if repos.is_installed(name, version) then
+         repos.delete_version(name, version, opts.deps_mode)
+      end
+      local dirs, err = prepare_install_dirs(name, version)
+      if not dirs then return nil, err end
+
+      local rollback = util.schedule_function(function()
+         fs.delete(path.install_dir(name, version))
+         fs.remove_dir_if_empty(path.versions_dir(name))
+      end)
    end
-   local dirs, err = prepare_install_dirs(name, version)
-   if not dirs then return nil, err end
-   
-   local rollback = util.schedule_function(function()
-      fs.delete(path.install_dir(name, version))
-      fs.remove_dir_if_empty(path.versions_dir(name))
-   end)
 
    ok, err = build.apply_patches(rockspec)
    if not ok then return nil, err end
@@ -405,6 +408,14 @@ function build.build_rockspec(rockspec, opts)
    ok, err = run_build_driver(rockspec)
    if not ok then return nil, err end
    
+   if opts.no_install then
+      fs.pop_dir()
+      if opts.need_to_fetch then
+         fs.pop_dir()
+      end
+      return name, version
+   end
+
    ok, err = install_files(rockspec, dirs)
    if not ok then return nil, err end
